Move Mock types closer to tests
